requirements.txt
# Core dependencies
python-docx==1.1.0
lxml==5.1.0

# API REST
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# CLI
click==8.1.7
tqdm==4.66.1

# Utilities
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0